feat(routes): integrates Product routes
